+++ include/pm3_cmd.h
@@ -605,6 +605,7 @@ typedef struct {
 
 #define CMD_HF_ISO14443A_SNIFF                                            0x0383
 #define CMD_HF_ISO14443A_SIMULATE                                         0x0384
+#define CMD_HF_ISO14443A_SIM_AID                                          0x1420
 
 #define CMD_HF_ISO14443A_READER                                           0x0385
 
@@ -775,6 +776,7 @@ typedef struct {
 #define FLAG_FORCED_ATQA        0x800
 #define FLAG_FORCED_SAK         0x1000
 #define FLAG_CVE21_0430         0x2000
+#define RATS_IN_DATA            0x3000
 
 
 #define MODE_SIM_CSN        0